requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.58.1
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.6
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3
